Attempt at not dropping FK index when filtered index is added
Main problem here is that the core core assumes that two unnamed indexes over an ordered set of properties are the same. This means calling HasIndex for won't even create the duplicate index. But if one of the indexes is filtered, then they are not equivalent. But filtering is a relational concept, so...